重新设计数据结构和UI 增加胸部和乳腺不同部位的数据

+++ src/main/java/com/zivy009/demo/springbootshirodwz/services/patient/service/IPatientService.java
@@ -14,6 +14,6 @@ public interface IPatientService extends IService<Patient> {
   int save(PatientDto patientDto);
   int delLogical(Long id);
   int del(Long id);
-  PatientDto selectById(Integer id);
+  PatientDto selectById(Long id);
   int update(PatientDto patientDto);
 }
